+++ public/js/tmsm-aquatonic-course-booking-public.js
@@ -752,12 +752,12 @@ var TmsmAquatonicCourseApp = TmsmAquatonicCourseApp || {};
     });
   }
 
-  $('.tmsm-aquatonic-course-birthdate input').mask("99/99/9999", {placeholder: TmsmAquatonicCourseApp.i18n.birthdateformat});
+  //$('.tmsm-aquatonic-course-birthdate input').mask("99/99/9999", {placeholder: TmsmAquatonicCourseApp.i18n.birthdateformat});
   //$('.tmsm-aquatonic-course-phone input').mask('+99-9999999999', {placeholder: ''});
-  $('.tmsm-aquatonic-course-phone input').mask('YZ0000000000', {placeholder: '__________', translation:  {
+  /*$('.tmsm-aquatonic-course-phone input').mask('YZ0000000000', {placeholder: '__________', translation:  {
     'Y': {pattern: /\+/, optional: true},
     'Z': {pattern: /[0-9]/, optional: true}
-  }});
+  }});*/
 
 
   /**