+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Booking Form: Inverse the 2 steps how to: reorder fields, conditional logic (participants not empty) on continue button, remove conditional logic on submit button
+
 ### 1.7.4: July 8th, 2021
 * Booking form validation refactoring
 * Gravity Forms 2.5 compatibility: remove condition on submit button, update frontend optimizations